requirements.txt
asgiref==3.7.2
autopep8==2.0.4
Django==4.2.6
djangorestframework @ git+https://github.com/encode/django-rest-framework@f56b85b7dd7e4f786e0769bba6b7609d4507da83
gunicorn==21.2.0
mysqlclient==2.2.0
packaging==23.2
Pillow==10.1.0
pycodestyle==2.11.1
sqlparse==0.4.4
whitenoise==6.6.0
